refactor(amazonq): Removing unwanted /dev and /doc code (#7760)
## Problem
- There is lot of duplicate and unwanted redundant code in
[aws-toolkit-vscode](https://github.com/aws/aws-toolkit-vscode)
repository.

## Solution

- This is the 2nd PR to remove unwanted code.
- Here is the 1st PR:
https://github.com/aws/aws-toolkit-vscode/pull/7735
---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.
