--- packages/core/src/amazonqDoc/app.ts
@@ -1,103 +1,0 @@
-/*!
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import * as vscode from 'vscode'
-import { ChatControllerEventEmitters, DocController } from './controllers/chat/controller'
-import { AmazonQAppInitContext } from '../amazonq/apps/initContext'
-import { MessageListener } from '../amazonq/messages/messageListener'
-import { fromQueryToParameters } from '../shared/utilities/uriUtils'
-import { getLogger } from '../shared/logger/logger'
-import { AuthUtil } from '../codewhisperer/util/authUtil'
-import { debounce } from 'lodash'
-import { DocChatSessionStorage } from './storages/chatSession'
-import { UIMessageListener } from './views/actions/uiMessageListener'
-import globals from '../shared/extensionGlobals'
-import { AppToWebViewMessageDispatcher } from '../amazonq/commons/connector/connectorMessages'
-import { docChat, docScheme } from './constants'
-import { TabIdNotFoundError } from '../amazonqFeatureDev/errors'
-import { DocMessenger } from './messenger'
-
-export function init(appContext: AmazonQAppInitContext) {
-    const docChatControllerEventEmitters: ChatControllerEventEmitters = {
-        processHumanChatMessage: new vscode.EventEmitter<any>(),
-        followUpClicked: new vscode.EventEmitter<any>(),
-        openDiff: new vscode.EventEmitter<any>(),
-        processChatItemVotedMessage: new vscode.EventEmitter<any>(),
-        stopResponse: new vscode.EventEmitter<any>(),
-        tabOpened: new vscode.EventEmitter<any>(),
-        processChatItemFeedbackMessage: new vscode.EventEmitter<any>(),
-        tabClosed: new vscode.EventEmitter<any>(),
-        authClicked: new vscode.EventEmitter<any>(),
-        formActionClicked: new vscode.EventEmitter<any>(),
-        processResponseBodyLinkClick: new vscode.EventEmitter<any>(),
-        insertCodeAtPositionClicked: new vscode.EventEmitter<any>(),
-        fileClicked: new vscode.EventEmitter<any>(),
-    }
-
-    const messenger = new DocMessenger(
-        new AppToWebViewMessageDispatcher(appContext.getAppsToWebViewMessagePublisher()),
-        docChat
-    )
-    const sessionStorage = new DocChatSessionStorage(messenger)
-
-    new DocController(
-        docChatControllerEventEmitters,
-        messenger,
-        sessionStorage,
-        appContext.onDidChangeAmazonQVisibility.event
-    )
-
-    const docProvider = new (class implements vscode.TextDocumentContentProvider {
-        async provideTextDocumentContent(uri: vscode.Uri): Promise<string> {
-            const params = fromQueryToParameters(uri.query)
-
-            const tabID = params.get('tabID')
-            if (!tabID) {
-                getLogger().error(`Unable to find tabID from ${uri.toString()}`)
-                throw new TabIdNotFoundError()
-            }
-
-            const session = await sessionStorage.getSession(tabID)
-            const content = await session.config.fs.readFile(uri)
-            const decodedContent = new TextDecoder().decode(content)
-            return decodedContent
-        }
-    })()
-
-    const textDocumentProvider = vscode.workspace.registerTextDocumentContentProvider(docScheme, docProvider)
-
-    globals.context.subscriptions.push(textDocumentProvider)
-
-    const docChatUIInputEventEmitter = new vscode.EventEmitter<any>()
-
-    new UIMessageListener({
-        chatControllerEventEmitters: docChatControllerEventEmitters,
-        webViewMessageListener: new MessageListener<any>(docChatUIInputEventEmitter),
-    })
-
-    const debouncedEvent = debounce(async () => {
-        const authenticated = (await AuthUtil.instance.getChatAuthState()).amazonQ === 'connected'
-        let authenticatingSessionIDs: string[] = []
-        if (authenticated) {
-            const authenticatingSessions = sessionStorage.getAuthenticatingSessions()
-
-            authenticatingSessionIDs = authenticatingSessions.map((session: any) => session.tabID)
-
-            // We've already authenticated these sessions
-            for (const session of authenticatingSessions) {
-                session.isAuthenticating = false
-            }
-        }
-
-        messenger.sendAuthenticationUpdate(authenticated, authenticatingSessionIDs)
-    }, 500)
-
-    AuthUtil.instance.secondaryAuth.onDidChangeActiveConnection(() => {
-        return debouncedEvent()
-    })
-    AuthUtil.instance.regionProfileManager.onDidChangeRegionProfile(() => {
-        return debouncedEvent()
-    })
-}
