--- packages/core/src/amazonqDoc/index.ts
@@ -1,10 +1,0 @@
-/*!
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export * from './types'
-export * from './session/sessionState'
-export * from './constants'
-export { Session } from './session/session'
-export { ChatControllerEventEmitters, DocController } from './controllers/chat/controller'
