--- packages/core/src/amazonqFeatureDev/controllers/chat/messenger/constants.ts
@@ -1,6 +1,0 @@
-/*!
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export type MessengerTypes = 'answer' | 'answer-part' | 'answer-stream' | 'system-prompt'
