--- packages/core/src/amazonqFeatureDev/index.ts
@@ -1,15 +1,0 @@
-/*!
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export * from './userFacingText'
-export * from './errors'
-export * from './session/sessionState'
-export * from './constants'
-export { Session } from './session/session'
-export { FeatureDevClient } from './client/featureDev'
-export { FeatureDevChatSessionStorage } from './storages/chatSession'
-export { TelemetryHelper } from '../amazonq/util/telemetryHelper'
-export { prepareRepoData, PrepareRepoDataOptions } from '../amazonq/util/files'
-export { ChatControllerEventEmitters, FeatureDevController } from './controllers/chat/controller'
