--- packages/core/src/amazonqFeatureDev/models.ts
@@ -1,12 +1,0 @@
-/*!
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export interface IManifestFile {
-    pomArtifactId: string
-    pomFolderName: string
-    hilCapability: string
-    pomGroupId: string
-    sourcePomVersion: string
-}
