+++ packages/core/src/shared/constants.ts
@@ -4,6 +4,7 @@
  */
 
 import * as vscode from 'vscode'
+import { manageAccessGuideURL } from '../amazonq/webview/ui/texts/constants'
 
 export const profileSettingKey = 'profile'
 export const productName: string = 'aws-toolkit-vscode'
@@ -196,3 +197,11 @@ export const amazonQVscodeMarketplace =
 export const crashMonitoringDirName = 'crashMonitoring'
 
 export const amazonQTabSuffix = '(Generated by Amazon Q)'
+
+/**
+ * Common strings used throughout the application
+ */
+
+export const uploadCodeError = `I'm sorry, I couldn't upload your workspace artifacts to Amazon S3 to help you with this task. You might need to allow access to the S3 bucket. For more information, see the [Amazon Q documentation](${manageAccessGuideURL}) or contact your network or organization administrator.`
+
+export const featureName = 'Amazon Q Developer Agent for software development'
