+++ packages/core/src/test/index.ts
@@ -22,6 +22,5 @@ export { getTestWorkspaceFolder } from '../testInteg/integrationTestsUtilities'
 export * from './codewhisperer/testUtil'
 export * from './credentials/testUtil'
 export * from './testUtil'
-export * from './amazonq/utils'
 export * from './fake/mockFeatureConfigData'
 export * from './shared/ui/testUtils'
